+++ packages/runtime/test/src/interpreter/opcode-list.ts
@@ -13,10 +13,11 @@ import {
   AppOptedIn, Arg, Assert, Balance, BitwiseAnd, BitwiseNot, BitwiseOr,
   BitwiseXor, Branch, BranchIfNotZero, BranchIfZero, Btoi,
   Byte, Bytec, Bytecblock, Concat, Div, Dup, Dup2, Ed25519verify,
-  EqualTo, Err, GetAssetDef, GetAssetHolding, Global,
+  EqualTo, Err, GetAssetDef, GetAssetHolding, GetBit, Global,
   GreaterThan, GreaterThanEqualTo, Gtxn, Gtxna, Int, Intc,
   Intcblock, Itob, Keccak256, Label, Len, LessThan, LessThanEqualTo,
   Load, Mod, Mul, Mulw, Not, NotEqualTo, Or, Pragma, PushBytes, PushInt, Return,
+  SetBit,
   Sha256, Sha512_256, Store, Sub, Substring, Substring3, Swap, Txn, Txna
 } from "../../../src/interpreter/opcode-list";
 import { DEFAULT_STACK_ELEM, MAX_UINT8, MAX_UINT64, MaxTEALVersion, MIN_UINT8 } from "../../../src/lib/constants";
@@ -3347,6 +3348,262 @@ describe("Teal Opcodes", function () {
       expectRuntimeError(
         () => op.execute(stack),
         RUNTIME_ERRORS.TEAL.ASSERT_STACK_LENGTH
+      );
+    });
+  });
+
+  describe("SetBit", () => {
+    let stack: Stack<StackElem>;
+    this.beforeEach(() => { stack = new Stack<StackElem>(); });
+
+    it("should set bit for uint64", () => {
+      const op = new SetBit([], 0);
+      stack.push(0n); // target
+      stack.push(4n); // index
+      stack.push(1n); // bit
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.equal(stack.pop(), 16n);
+
+      stack.push(16n);
+      stack.push(0n);
+      stack.push(1n);
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.equal(stack.pop(), 17n);
+
+      stack.push(15n);
+      stack.push(0n);
+      stack.push(0n);
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.equal(stack.pop(), 14n);
+
+      stack.push(0n);
+      stack.push(63n);
+      stack.push(1n);
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.equal(stack.pop(), 2n ** 63n);
+
+      stack.push(MAX_UINT64);
+      stack.push(1n);
+      stack.push(0n);
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.equal(stack.pop(), MAX_UINT64 - 2n);
+    });
+
+    it("should panic if index bit is not uint64", () => {
+      const op = new SetBit([], 0);
+      stack.push(0n); // target
+      stack.push(new Uint8Array([1, 2])); // index
+      stack.push(1n); // bit
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.INVALID_TYPE
+      );
+    });
+
+    it("should panic if set bit is not uint64", () => {
+      const op = new SetBit([], 0);
+      stack.push(0n); // target
+      stack.push(4n); // index
+      stack.push(new Uint8Array([1, 2])); // bit
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.INVALID_TYPE
+      );
+    });
+
+    it("should panic if stack length is less than 3", () => {
+      const op = new SetBit([], 0);
+      stack.push(0n);
+      stack.push(4n);
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.ASSERT_STACK_LENGTH
+      );
+    });
+
+    it("should panic if set bit is greater than 1", () => {
+      const op = new SetBit([], 0);
+      stack.push(0n);
+      stack.push(4n);
+      stack.push(20n);
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.SET_BIT_VALUE_ERROR
+      );
+    });
+
+    it("should panic if set bit index is greater than 63 and target is uint64", () => {
+      const op = new SetBit([], 0);
+      stack.push(0n);
+      stack.push(400n);
+      stack.push(1n);
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.SET_BIT_INDEX_ERROR
+      );
+    });
+
+    it("should set bit in bytes array", () => {
+      const op = new SetBit([], 0);
+      stack.push(new Uint8Array([0, 0, 0])); // target
+      stack.push(8n); // index
+      stack.push(1n); // bit
+
+      // set 8 th bit of bytes to 1 i.e 8th bit will be highest order bit of second byte
+      // so second byte will become 2 ** 7 = 128
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.deepEqual(stack.pop(), new Uint8Array([0, 2 ** 7, 0]));
+
+      // set bit again to 0
+      stack.push(new Uint8Array([0, 2 ** 7, 0])); // target
+      stack.push(8n); // index
+      stack.push(0n); // bit
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.deepEqual(stack.pop(), new Uint8Array([0, 0, 0]));
+
+      stack.push(new Uint8Array([0, 2 ** 7, 0])); // target
+      stack.push(0n); // index
+      stack.push(1n); // bit
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.deepEqual(stack.pop(), new Uint8Array([2 ** 7, 2 ** 7, 0]));
+
+      stack.push(new Uint8Array([0, 2 ** 7, 0])); // target
+      stack.push(7n); // index
+      stack.push(1n); // bit
+
+      op.execute(stack);
+
+      assert.equal(stack.length(), 1);
+      assert.deepEqual(stack.pop(), new Uint8Array([1, 2 ** 7, 0]));
+    });
+
+    it("should panic if index bit in out of bytes array", () => {
+      const op = new SetBit([], 0);
+      stack.push(new Uint8Array([0, 0, 0])); // target
+      stack.push(80n); // index
+      stack.push(1n); // bit
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.SET_BIT_INDEX_BYTES_ERROR
+      );
+
+      stack.push(new Uint8Array(8).fill(0)); // target
+      stack.push(64n * 8n + 1n); // index
+      stack.push(1n); // bit
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.SET_BIT_INDEX_BYTES_ERROR
+      );
+    });
+  });
+
+  describe("GetBit", () => {
+    let stack: Stack<StackElem>;
+    this.beforeEach(() => { stack = new Stack<StackElem>(); });
+
+    it("should push correct bit to stack(uint64)", () => {
+      const op = new GetBit([], 0);
+      stack.push(8n); // target
+      stack.push(3n); // index
+
+      op.execute(stack);
+      assert.equal(stack.pop(), 1n);
+
+      stack.push(8n); // target
+      stack.push(0n); // index
+
+      op.execute(stack);
+      assert.equal(stack.pop(), 0n);
+    });
+
+    it("should push correct bit to stack(bytes array)", () => {
+      const op = new GetBit([], 0);
+      stack.push(new Uint8Array([0, 128, 1])); // target
+      stack.push(8n); // index
+      op.execute(stack);
+      assert.equal(stack.pop(), 1n);
+
+      stack.push(new Uint8Array([1, 4, 1])); // target
+      stack.push(23n); // index
+      op.execute(stack);
+      assert.equal(stack.pop(), 1n);
+
+      stack.push(new Uint8Array([4, 0, 1])); // target
+      stack.push(6n); // index
+      op.execute(stack);
+      assert.equal(stack.pop(), 0n);
+    });
+
+    it("should panic if stack length is less than 2", () => {
+      const op = new GetBit([], 0);
+      stack.push(0n);
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.ASSERT_STACK_LENGTH
+      );
+    });
+
+    it("should panic if index bit is not uint64", () => {
+      const op = new GetBit([], 0);
+      stack.push(8n); // target
+      stack.push(new Uint8Array(0)); // index
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.INVALID_TYPE
+      );
+    });
+
+    it("should panic if index bit in out of uint64 bits", () => {
+      const op = new GetBit([], 0);
+      stack.push(8n); // target
+      stack.push(500n); // index
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.SET_BIT_INDEX_ERROR
+      );
+    });
+
+    it("should panic if index bit in out of bytes array", () => {
+      const op = new GetBit([], 0);
+      stack.push(new Uint8Array(0)); // target
+      stack.push(500n); // index
+
+      expectRuntimeError(
+        () => op.execute(stack),
+        RUNTIME_ERRORS.TEAL.SET_BIT_INDEX_BYTES_ERROR
       );
     });
   });