+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 + Update runtime to process execParams.deployASA, deploySSC, OptInToASA, OptIntoSSC
 + Exported `@algorand-builder/algob`, `@algorand-builder/runtime` error types and make it accessible for API documentation.
 + Added `debugStack` option in `runtime.executeTx()` to print stack (upto depth = debugStack) after each opcode execution.
++ Tealv3 support in `@algo-builder/runtime`: `getBit`, `setBit`
 
 ### Bug fixes